+++ ecom_nov/src/main/java/com/ecom/mainapp/services/ProductService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface ProductService {
     List<Product> getProducts();
-    Product getProductById(int id) throws ProductNotFoundException;
+    Product getProductById(long id) throws ProductNotFoundException;
     Product updateProduct(Product product) throws ProductNotFoundException;
-    boolean deleteProductById(int id) throws ProductNotFoundException;
+    boolean deleteProductById(long id) throws ProductNotFoundException;
     Product createProduct(Product product);
 }
